Fix inconsistent type in the cube creation code
I have to be honest I am not sure how that ever worked! This code
reorders the cube from the Avogadro ordering to the VTK ordering, but it
was incorrectly casting to double* rather than float*. This was causing
a segfault in my testing, once fixed it works as expected.

+++ avogadro/vtk/vtkglwidget.cpp
@@ -50,7 +50,7 @@ vtkImageData* createCubeImageData(Core::Cube* cube)
 
   data->AllocateScalars(VTK_FLOAT, 1);
 
-  auto* dataPtr = static_cast<double*>(data->GetScalarPointer());
+  auto* dataPtr = static_cast<float*>(data->GetScalarPointer());
   std::vector<float>* cubePtr = cube->data();
 
   for (int i = 0; i < dim.x(); ++i) {
